requirements.txt
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-ads1x15
requests
python-dotenv
